Remove extensive round-trip consistency checks and visualizations from monogenic curvelet example
- Eliminated detailed component comparisons and frequency domain analyses to streamline the example.
- Removed print statements and visualizations related to scalar and Riesz component comparisons.
- Simplified the overall structure of the example for better clarity and focus on core functionalities.
